restructure about content into subpages
