+++ dp-game-webapp/app/about/components.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+import clsx from "clsx";
+import { Switch } from "@headlessui/react";
+
+export function Toggle({
+  enabled,
+  setEnabled,
+}: {
+  enabled: boolean;
+  setEnabled: (value: boolean) => void;
+}) {
+  return (
+    <Switch
+      checked={enabled}
+      onChange={setEnabled}
+      className={clsx(
+        "relative inline-flex h-6 w-11 flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition-colors duration-200 ease-in-out focus:outline-none focus:ring-2 focus:ring-indigo-600 focus:ring-offset-2",
+        enabled ? "bg-indigo-600" : "bg-gray-200",
+      )}
+    >
+      <span className="sr-only">Use setting</span>
+      <span
+        aria-hidden="true"
+        className={clsx(
+          "pointer-events-none inline-block h-5 w-5 transform rounded-full bg-white shadow ring-0 transition duration-200 ease-in-out",
+          enabled ? "translate-x-5" : "translate-x-0",
+        )}
+      />
+    </Switch>
+  );
+}
